pyproject.toml
[project]
name = "tuican"
version = "0.1.0"
authors = [
    { name="G0-G4"},
]
description = "telegram ui components to build app alike bots"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.4",
]

[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project.urls]
Homepage = "https://github.com/G0-G4/TUIcan/tree/main"